Bluetooth enabe request on build app
